requirements.txt
django-model-utils>=2.0
django-appconf
python-dateutil
djangorestframework
psycopg2
